backend/database_dynamo/pyproject.toml
[project]
name = "alex-database-dynamo"
version = "0.1.0"
description = "DynamoDB database package for Alex Financial Planner"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.8",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "moto[dynamodb]>=4.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
